+++ src/Path.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace CoRex\Filesystem;
 
 use CoRex\Helpers\Arr;
@@ -9,10 +11,10 @@ class Path
     /**
      * Get path to root of site.
      *
-     * @param string|array $segments Dot notation is supported in string. Default null.
+     * @param string|string[] $segments Dot notation is supported in string. Default null.
      * @return string
      */
-    public static function root($segments = null)
+    public static function root($segments = null): string
     {
         $path = __DIR__;
         for ($c1 = 0; $c1 < 4; $c1++) {
@@ -26,10 +28,10 @@ class Path
     /**
      * Get path to current package.
      *
-     * @param string|array $segments Dot notation is supported in string. Default null.
+     * @param string|string[] $segments Dot notation is supported in string. Default null.
      * @return string
      */
-    public static function packageCurrent($segments = null)
+    public static function packageCurrent($segments = null): string
     {
         return self::package(null, null, $segments);
     }
@@ -40,10 +42,10 @@ class Path
      *
      * @param string $vendor Default null which means current.
      * @param string $package Default null which means current.
-     * @param @param string|array $segments Dot notation is supported in string. Default null.
+     * @param string|string[] $segments Dot notation is supported in string. Default null.
      * @return string
      */
-    public static function package($vendor = null, $package = null, $segments = null)
+    public static function package(?string $vendor = null, ?string $package = null, $segments = null): string
     {
         $path = dirname(dirname(static::packagePath()));
         if ($package === null) {
@@ -62,7 +64,7 @@ class Path
      *
      * @return string
      */
-    public static function vendorName()
+    public static function vendorName(): string
     {
         $path = static::packagePath();
         return basename(dirname($path));
@@ -73,7 +75,7 @@ class Path
      *
      * @return string
      */
-    public static function packageName()
+    public static function packageName(): string
     {
         $path = static::packagePath();
         return basename($path);
@@ -85,7 +87,7 @@ class Path
      *
      * @return string
      */
-    protected static function packagePath()
+    protected static function packagePath(): string
     {
         return dirname(__DIR__);
     }
@@ -94,10 +96,10 @@ class Path
      * Add segments to path.
      *
      * @param string $path
-     * @param @param string|array $segments Dot notation is supported in string.
+     * @param string|string[] $segments Dot notation is supported in string.
      * @return string
      */
-    private static function addSegmentsToPath($path, $segments)
+    private static function addSegmentsToPath(string $path, $segments): string
     {
         $segments = Arr::toArray($segments);
         if (count($segments) > 0) {