+++ tests/JsonTest.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\CoRex\Filesystem;
 
 use CoRex\Filesystem\Directory;
@@ -10,16 +12,19 @@ use PHPUnit\Framework\TestCase;
 
 class JsonTest extends TestCase
 {
+    /** @var string */
     private $tempDirectory;
+
+    /** @var string */
     private $tempFilename;
 
     /**
-     * Test constructor.
+     * Test.
      *
      * @throws \ReflectionException
      * @throws \Exception
      */
-    public function testConstructor()
+    public function testConstructor(): void
     {
         $json = new Json($this->tempFilename);
         $this->assertEquals($this->tempFilename, Obj::getProperty('filename', $json));
@@ -32,7 +37,7 @@ class JsonTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testGetFilename()
+    public function testGetFilename(): void
     {
         $json = new Json($this->tempFilename);
         $this->assertEquals($this->tempFilename, $json->getFilename());
@@ -43,7 +48,7 @@ class JsonTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testExist()
+    public function testExist(): void
     {
         // Make sure it is removed.
         File::delete($this->tempFilename);
@@ -59,10 +64,10 @@ class JsonTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testKeyOrderValueOneToTwo()
+    public function testKeyOrderValueOneToTwo(): void
     {
-        $value1 = md5(mt_rand(1, 100000)) . '1';
-        $value2 = md5(mt_rand(1, 100000)) . '2';
+        $value1 = md5((string)mt_rand(1, 100000)) . '1';
+        $value2 = md5((string)mt_rand(1, 100000)) . '2';
         $json = new Json($this->tempFilename, ['value1', 'value2']);
         $json->set('value1', $value1);
         $json->set('value2', $value2);
@@ -79,10 +84,10 @@ class JsonTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testKeyOrderValueTwoToOne()
+    public function testKeyOrderValueTwoToOne(): void
     {
-        $value1 = md5(mt_rand(1, 100000)) . '1';
-        $value2 = md5(mt_rand(1, 100000)) . '2';
+        $value1 = md5((string)mt_rand(1, 100000)) . '1';
+        $value2 = md5((string)mt_rand(1, 100000)) . '2';
         $json = new Json($this->tempFilename, ['value2', 'value1']);
         $json->set('value1', $value1);
         $json->set('value2', $value2);
@@ -99,10 +104,10 @@ class JsonTest extends TestCase
      *
      * @throws \Exception
      */
-    public function testSave()
+    public function testSave(): void
     {
-        $value1 = md5(mt_rand(1, 100000)) . '1';
-        $value2 = md5(mt_rand(1, 100000)) . '2';
+        $value1 = md5((string)mt_rand(1, 100000)) . '1';
+        $value2 = md5((string)mt_rand(1, 100000)) . '2';
         $json = new Json($this->tempFilename);
         $json->set('value1', $value1);
         $json->set('value2', $value2);
@@ -117,7 +122,7 @@ class JsonTest extends TestCase
     /**
      * Setup.
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->tempDirectory = sys_get_temp_dir();
@@ -129,7 +134,7 @@ class JsonTest extends TestCase
     /**
      * Tear down.
      */
-    protected function tearDown()
+    protected function tearDown(): void
     {
         parent::tearDown();
         Directory::delete($this->tempDirectory);