+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+## 2.0.0
+
+### Changed
+- Require php 7.2+
+- Updated code to comply with Coding Standard.
+- Changed File::type() to return '' instead of false when failing.
+- Changed File::mimetype() to return '' instead of false when failing.
+
 ## 1.2.1
 
 ### Fixed